[HeterogeneousDWARF] Avoid pathological slowdown in PruningFunctionCloner
The ValueMapper used in the PruningFunctionCloner becomes significantly
slower when support for module level metadata changes is enabled.

Instead, use the ValueMaterializer interface to handle the cloning
of non-global DILifetimes.

There is potential for some cleanup here, and unification with the
approach used in the non-pruning clone methods, but this patch fixes an
urgent performance regression in a common case.

Change-Id: I5c30b84fa8af4bd92496a06d4fe3f538cc82ef6f