[HeterogeneousDWARF] Add pass to strip heterogeneous debug info above -O0
Until we have audited all passes which run above -O0 and implemented
heterogeneous debug info handling in them it is very easy for an end
user to accidentally omit -O0 when compiling with any tool other
than clang and either get nonsense DWARF output or a cryptic compilation
failure (typically an unrelated assertion).

As it will still take some time before work on -O1 support even starts,
let alone -O2 and -O3, this patch adds HeterogeneousDebugVerify, a pass
which runs on every module early (pre ISel) to strip heterogeneous debug
info out of any module which is being compiled above -O0.

For development and testing at higher optimization levels the LLVM
option -disable-heterogeneous-debug-verify will disable this check.

Change-Id: I5e53b7caab9657367574ac6755f1144b4fe61c35

[NFC][HeterogeneousDebugVerify] Forward CodeGenOpt::Level instead of getting it from the TargetMachine

Calling TargetMachine::getOptLevel introduced a dependency between
LLVMCore and LLVMTarget. But LLVMTarget already depends on LLVMCore.

When configuring the project with -DBUILD_SHARED_LIBS=ON, this
dependency resulted in an undefined reference.

Change-Id: I0e7b612edc2b49592f1f0e6b3063d065b5e66e01

[HeterogeneousDebugVerify] Keep the DILocalVariables and
DIGlobalVariables after optimization

Even if locals/globals cannot be mapped back to a location, the compiler
keeps these entities in the debug-info.

Change-Id: I30e5ab5d036915b48cd9ab17a8a9aae8022cc551

[NFC][HeterogeneousDwarf] Iterate only over dbg.def/kill intrinsics

Iterate only over debug intrinsics instead of over all the instructions
in the module.

This should improve the execution time of this pass.

Change-Id: Ie5e1b3208803b4dc80d7fa5038d82d759d3f260e

[HeterogeneousDwarf] Stop printing a warning in HeterogeneousDebugVerify

In HeterogeneousDebugVerify we emitted a warning to inform that we
were degrading the debug information when optimizations are enabled.

When compiling with -Werror, this results in an error.

Change-Id: Ia5a87e1de433386f1a9bdb31f180905d98c697bb