[HeterogeneousDWARF] Drop DW_AT_active_lane encoding and DW_AT_augmentation
A collision upstream means we need to update the DW_AT_active_lane and
DW_AT_augmentation encodings.

The DWARF committee has expressed concern about the augmentation
attribute, so we remove it entirely.

The DW_AT_active_lane has not been produced by the compiler yet, so we
move its encoding to TBA.

Change-Id: Id89af670c8c49be06ab8da552b0d9d22e04c84f1